requirements.txt
flask
flask-cors
openai-whisper
tensorflow
numpy
soundfile
